Create new file characterization.py for processing bias and current data into PlasmaPy characteristics
